Create Introduction-Business Problem
A description of the problem and a discussion of the background. (15 marks)
In this project, I will be working with data from New York City. 
I will be investigating where in New York City one should open a Swedish restaurant. 
Several restaurateurs are looking into publicizing Swedish cuisine in the United States.
The problem is that in NYC, cost of operation is extremely high and in order to succeed, 
a restaurant must be located in an area with ample customers and low competition.