Create Data
A description of the data and how it will be used to solve the problem. (15 marks)
I will be using data about the average cost per square foot of real estate throughout each of the 5 boroughs: 
Manhattan, The Bronx, Brooklyn, Queens, Staten Island. 
I will also look into the  specific competition of my restaurant as well as the foot traffic in the areas to 
find the ideal location for my restaurant.